Move config to app root
