+++ app/clients/db/session.py
@@ -1,7 +1,7 @@
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-from app.core import config
+from app import config
 
 engine = create_engine(
     config.SQLALCHEMY_DATABASE_URI,
